+++ changelog.md
@@ -18,6 +18,8 @@ Codion Change Log
 ### is.codion.common.i18n
 - Messages, mnemonic bounds checking added.
 - Messages, icelandic Clear mnemonic fixed.
+### is.codion.common.rmi
+- SerializationFilter, minor refactoring.
 ### is.codion.dbms.h2
 - H2Database, security validation for the H2 database script injection.
 ### is.codion.swing.common.ui